feat: Refactor username label component to handle gif URL updates
The username label component has been refactored to handle updates to the gif URL. The `getUrlLabel` method has been added to retrieve the URL based on the user ID. This method is called during component initialization and also when the `currentGifUrl` observable emits a new value. This ensures that the gif URL is always up to date.